Added unencrypted namespace file_service
