+++ lib/porky_lib.rb
@@ -5,4 +5,6 @@ module PorkyLib
   require 'porky_lib/file_service'
   require 'porky_lib/symmetric'
   require 'porky_lib/version'
+  require 'porky_lib/unencrypted'
+  require 'porky_lib/file_service_helper'
 end
