+++ lib/porky_lib/file_service.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
-require 'aws-sdk-s3'
 require 'singleton'
+require 'porky_lib/file_service_helper'
 
 class PorkyLib::FileService
   include Singleton
+  include FileServiceHelper
 
   class FileServiceError < StandardError; end
   class FileSizeTooLargeError < StandardError; end
@@ -61,11 +62,7 @@ class PorkyLib::FileService
     data = file_data(file)
     file_key = options.key?(:directory) ? "#{options[:directory]}/#{SecureRandom.uuid}" : SecureRandom.uuid
 
-    tempfile = if options[:store_raw_file]
-                 write_tempfile(data, file_key)
-               else
-                 encrypt_file_contents(data, key_id, file_key, options)
-               end
+    tempfile = encrypt_file_contents(data, key_id, file_key, options)
 
     begin
       perform_upload(bucket_name, file_key, tempfile, options)
@@ -119,40 +116,6 @@ class PorkyLib::FileService
 
   private
 
-  def input_invalid?(file, bucket_name, key_id)
-    file.nil? || bucket_name.nil? || key_id.nil?
-  end
-
-  def file_size_invalid?(file)
-    file.bytesize > max_size || (File.file?(file) && File.size(file) > max_size)
-  end
-
-  def file_data(file)
-    File.file?(file) ? File.read(file) : file
-  end
-
-  def max_size
-    PorkyLib::Config.config[:max_file_size]
-  end
-
-  def max_file_size
-    {
-      B: 1024,
-      KB: 1024 * 1024,
-      MB: 1024 * 1024 * 1024,
-      GB: 1024 * 1024 * 1024 * 1024
-    }.each_pair { |symbol, bytes| return "#{(max_size.to_f / (bytes / 1024)).round(2)}#{symbol}" if max_size < bytes }
-  end
-
-  def perform_upload(bucket_name, file_key, tempfile, options)
-    obj = s3.bucket(bucket_name).object(file_key)
-    if options.key?(:metadata)
-      obj.upload_file(tempfile.path, metadata: options[:metadata])
-    else
-      obj.upload_file(tempfile.path)
-    end
-  end
-
   def decrypt_file_contents(tempfile)
     file_contents = tempfile.read
 
@@ -189,23 +152,11 @@ class PorkyLib::FileService
     }.to_json
   end
 
-  def write_tempfile(file_contents, file_key)
-    tempfile = Tempfile.new(file_key)
-    tempfile << file_contents
-    tempfile.close
-
-    tempfile
-  end
-
   def presign_url_expires_in
     PorkyLib::Config.config[:presign_url_expires_in]
   end
 
   def s3_client
     @s3_client ||= Aws::S3::Client.new
-  end
-
-  def s3
-    @s3 ||= Aws::S3::Resource.new
   end
 end
