+++ lib/porky_lib/file_service_helper.rb
@@ -1,0 +1,51 @@
+# frozen_string_literal: true
+
+require 'aws-sdk-s3'
+
+module FileServiceHelper
+  def input_invalid?(file, bucket_name, key_id)
+    file.nil? || bucket_name.nil? || key_id.nil?
+  end
+
+  def file_size_invalid?(file)
+    file.bytesize > max_size || (File.file?(file) && File.size(file) > max_size)
+  end
+
+  def file_data(file)
+    File.file?(file) ? File.read(file) : file
+  end
+
+  def write_tempfile(file_contents, file_key)
+    tempfile = Tempfile.new(file_key)
+    tempfile << file_contents
+    tempfile.close
+
+    tempfile
+  end
+
+  def perform_upload(bucket_name, file_key, tempfile, options)
+    obj = s3.bucket(bucket_name).object(file_key)
+    if options.key?(:metadata)
+      obj.upload_file(tempfile.path, metadata: options[:metadata])
+    else
+      obj.upload_file(tempfile.path)
+    end
+  end
+
+  def s3
+    @s3 ||= Aws::S3::Resource.new
+  end
+
+  def max_size
+    PorkyLib::Config.config[:max_file_size]
+  end
+
+  def max_file_size
+    {
+      B: 1024,
+      KB: 1024 * 1024,
+      MB: 1024 * 1024 * 1024,
+      GB: 1024 * 1024 * 1024 * 1024
+    }.each_pair { |symbol, bytes| return "#{(max_size.to_f / (bytes / 1024)).round(2)}#{symbol}" if max_size < bytes }
+  end
+end
