+++ lib/porky_lib/unencrypted.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Unencrypted
+  require 'porky_lib/unencrypted/file_service'
+end
