+++ lib/porky_lib/unencrypted/file_service.rb
@@ -1,0 +1,47 @@
+# frozen_string_literal: true
+
+require 'singleton'
+require 'porky_lib/file_service_helper'
+
+class Unencrypted::FileService
+  include Singleton
+  include FileServiceHelper
+
+  class FileServiceError < StandardError; end
+  class FileSizeTooLargeError < StandardError; end
+
+  def read(bucket_name, file_key, options = {})
+    tempfile = Tempfile.new
+
+    begin
+      object = s3.bucket(bucket_name).object(file_key)
+      raise FileSizeTooLargeError, "File size is larger than maximum allowed size of #{max_file_size}" if object.content_length > max_size
+
+      object.download_file(tempfile.path, options)
+    rescue Aws::Errors::ServiceError => e
+      raise FileServiceError, "Attempt to download a file from S3 failed.\n#{e.message}"
+    end
+
+    tempfile
+  end
+
+  def write(file, bucket_name, key_id, options = {})
+    raise FileServiceError, 'Invalid input. One or more input values is nil' if input_invalid?(file, bucket_name, key_id)
+    raise FileSizeTooLargeError, "File size is larger than maximum allowed size of #{max_file_size}" if file_size_invalid?(file)
+
+    data = file_data(file)
+    file_key = options.key?(:directory) ? "#{options[:directory]}/#{SecureRandom.uuid}" : SecureRandom.uuid
+
+    tempfile = write_tempfile(data, file_key)
+
+    begin
+      perform_upload(bucket_name, file_key, tempfile, options)
+    rescue Aws::Errors::ServiceError => e
+      raise FileServiceError, "Attempt to upload a file to S3 failed.\n#{e.message}"
+    end
+
+    # Remove tempfile from disk
+    tempfile.unlink
+    file_key
+  end
+end
